auth/requirements.txt
bcrypt==4.0.1
dependency-injector==4.41.0
flask-smorest==0.41.0
Flask==2.2.3
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
gevent==22.10.2
gunicorn==20.1.0
psycopg2-binary==2.9.6
pydantic==1.10.7
PyJWT==2.6.0
redis==4.5.4
